+++ just_audio/test/just_audio_test.dart
@@ -1288,6 +1288,7 @@ final icyMetadata = IcyMetadata(
   ),
   info: IcyInfo(
     title: 'title',
+    artist: 'artist',
     url: 'url',
   ),
 );
@@ -1303,6 +1304,7 @@ final icyMetadataMessage = IcyMetadataMessage(
   ),
   info: IcyInfoMessage(
     title: 'title',
+    artist: 'artist',
     url: 'url',
   ),
 );
@@ -1516,6 +1518,7 @@ class MockAudioPlayer implements AudioPlayerPlatform {
         ),
         info: IcyInfoMessage(
           title: 'title',
+          artist: 'artist',
           url: url,
         ),
       ),